// src/renderers/webgl2-offscreen/backend/backend.types.ts
export enum BackendWorkerMessage {
    LOG = -1,
    INITIALIZE = 0,
    INITIALIZE_RESULT = 1,
    RESIZE = 2,
    RESIZE_RESULT = 3,
    ENABLE_IMAGE_BOUNDARY_MASK = 4,
    ENABLE_IMAGE_BOUNDARY_MASK_RESULT = 5,
    SET_BACKGROUND_COLOR = 6,
    SET_BACKGROUND_COLOR_RESULT = 7,
    SET_MASKS = 8,
    SET_MASKS_RESULT = 9,
    SET_SELECTION_MASK = 10,
    SET_SELECTION_MASK_RESULT = 11,
    SET_VIEW_TRANSFORM = 12,
    SET_VIEW_TRANSFORM_RESULT = 13,
    SET_LAYER_ORDER = 14,
    SET_LAYER_ORDER_RESULT = 15,
    QUEUE_CREATE_LAYER_PASSES = 16,
    QUEUE_CREATE_LAYER_PASSES_RESULT = 17,
    APPLY_SELECTION_MASK_TO_ALPHA_CHANNEL = 18,
    APPLY_SELECTION_MASK_TO_ALPHA_CHANNEL_RESULT = 19,
    TAKE_SNAPSHOT = 20,
    TAKE_SNAPSHOT_RESULT = 21,
    START_BRUSH_STROKE = 22,
    START_BRUSH_STROKE_RESULT = 23,
    MOVE_BRUSH_STROKE = 24,
    MOVE_BRUSH_STROKE_RESULT = 25,
    STOP_BRUSH_STROKE = 26,
    STOP_BRUSH_STROKE_RESULT = 27,
    CREATE_BRUSH_PREVIEW = 28,
    CREATE_BRUSH_PREVIEW_RESULT = 29,
    DISPOSE = 30,
    DISPOSE_RESULT = 31,
    SET_DIRTY = 32,
    REQUEST_FRONTEND_SVG = 33,
    REQUEST_FRONTEND_SVG_RESULT = 34,
    REQUEST_FRONTEND_TEXTURE = 35,
    REQUEST_FRONTEND_TEXTURE_RESULT = 35,
    CREATE_MESH_CONTROLLER = 36,
    CREATE_MESH_CONTROLLER_RESULT = 37,
    UPDATE_MESH_CONTROLLER = 38,
    UPDATE_MESH_CONTROLLER_RESULT = 39,
    DISPOSE_MESH_CONTROLLER = 40,
    DISPOSE_MESH_CONTROLLER_RESULT = 41,
}